Refactoring for support of modular parsers
